Assign sales rep to team on sales rep list
Assign sales rep to team on sales rep list
on sales rep list select the sales rep and click on assign to direct team.

Filter team set on sales rep list, also set the condition on team to sales rep list there is not display the filter team because that is already inside that team. team to sales rep list.