Location management Filter section
Location management Filter section on company admin display all the team, all the sales rep, and all the customer.
when sales manager login display his team and after select the team to get and load the sales rep and after select the sales rep to get and load the customer.